Renamed the Invoice and import_settings folder in the modules directory and updated all import paths where the folder is used.
Renamed the Invoice and import_settings folder in the modules directory and updated all import paths where the folder is used.
